+++ src/GeneFinder.jl
@@ -5,10 +5,10 @@ using TestItems: @testitem
 
 
 include("algorithms/simplefinder.jl")
-export ORF, simplefinder
+export ORF, CDS, Proteins, simplefinder, findcds, findproteins
 
-include("findgenes.jl")
-export CDS, Proteins, findcds, findproteins
+# include("findgenes.jl")
+
 
 # Write your package code here.