+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
-# [UNRELEASED](https://github.com/camilogarciabotero/GeneFinder.jl/compare/v0.0.1...main)
+## [UNRELEASED](https://github.com/camilogarciabotero/GeneFinder.jl/compare/v0.0.1...main)
+
+## [0.0.2]
+
+- Add `findcds` and `findproteins` function for the first algorithm
+- Add docs for `simplefinder` functions and structs
+- Add test using `TestItems` package
 
 ## [0.0.1]